+++ src/Configuration/ProvidesScriptVariables.php
@@ -5,6 +5,7 @@ namespace SleepingOwl\Admin\Configuration;
 use Illuminate\Support\Facades\Request;
 use SleepingOwl\Admin\Contracts\Template\TemplateInterface;
 use SleepingOwl\Admin\Traits\DatePicker;
+use SleepingOwl\Admin\Traits\MaxFileSizeTrait;
 
 /**
  * Trait ProvidesScriptVariables.
@@ -12,7 +13,7 @@ use SleepingOwl\Admin\Traits\DatePicker;
  */
 trait ProvidesScriptVariables
 {
-    use DatePicker;
+    use DatePicker, MaxFileSizeTrait;
 
     /**
      * Получение массива глобальных
@@ -28,6 +29,8 @@ trait ProvidesScriptVariables
             $lang = trans('sleeping_owl::lang', [], 'en');
         }
 
+        // $maxFileSize = $this->convertMB(ini_get('upload_max_filesize'));
+
         $state_filters = $this->config['state_filters'];
         if (! $this->config['state_datatables']) {
             $state_filters = false;
@@ -42,6 +45,7 @@ trait ProvidesScriptVariables
             'wysiwyg' => $this->config['wysiwyg'],
             'template' => $this->app[TemplateInterface::class]->toArray(),
             'user_id' => auth()->id(),
+            'max_file_size' => $this->getMaxFileSize(),
             'datetime_format' => $this->generatePickerFormat($this->config['datetimeFormat']),
             'date_format' => $this->generatePickerFormat($this->config['dateFormat']),
             'state_datatables' => $this->config['state_datatables'],